+++ FlightsSystem.Core/DAL/IBasicDB.cs
@@ -4,7 +4,7 @@ namespace FlightsSystem.Core.DAL
 {
     public interface IBasicDB<T> where T : IPoco
     {
-        T Get(int id);
+        T Get(long id);
         IList<T> GetAll();
         void Add(T t);
         void Remove(T t);
