+++ FlightsSystem.Core/DAL/IFlightDAO.cs
@@ -6,7 +6,7 @@ namespace FlightsSystem.Core.DAL
     public interface IFlightDAO : IBasicDB<Flight>
     {
         Dictionary<Flight, int> GetAllFlightsVacancy();
-        Flight GetFlightById(int id);
+        Flight GetFlightById(long id);
         IList<Flight> GetFlightsByCustomer(Customer customer);
         IList<Flight> GetFlightsByDepartureDate(DateTime departureDate);
         IList<Flight> GetFlightsByDestinationCountry(Country destinationCountry);
